+++ CHANGELOG.md
@@ -22,6 +22,10 @@ UNRELEASED
 ### Changed
 - All `src` scripts now exit with success after being called with `--help`.
 - Minor optimizations on `contrib` scripts.
+- Huge speed improvement on `cpt-export`
+
+### Fixed
+- Fixed `cpt-cat` not using the `CPT_ROOT` value.
 
 ### Removed
 - Removed C binaries `cpt-readlink` and `cpt-stat` and instead replaced them